% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{my_data}
\alias{my_data}
\title{Sample data from the Windsor API.}
\format{
A data frame with 1677 rows and 6 variables:
\describe{
\item{data.campaign}{name of the campaign}
\item{data.clicks}{number of clicks}
\item{data.spend}{spend data}
\item{data.medium}{the type of referals}
\item{data.source}{source of the data (google, facebook, etc.)}
\item{googlesheets}{googlesheets id}
}
}
\source{
\url{https://www.windsor.ai/}
}
\usage{
my_data
}
\description{
A dataset containing sample gooole and facebook ads data
fetched from windsor.ai API. See more at: https://www.windsor.ai/api-fields/
}
\keyword{datasets}
