% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windsor_fetch_googleAds.R
\name{windsor_fetch_googleAds}
\alias{windsor_fetch_googleAds}
\title{Windsor fetch Google Ads
Fetch data from google Ads via windsor.ai API}
\usage{
windsor_fetch_googleAds(
  api_key,
  date_preset = "last_7d",
  fields = c("source", "campaign", "clicks", "medium", "sessions", "spend")
)
}
\arguments{
\item{api_key}{Your api key to access Windsor.ai API}

\item{date_preset}{the period for which data is fetched from the API.
See https://www.windsor.ai/api-fields/ for details}

\item{fields}{The fields fetched from the API for a given connector
See https://www.windsor.ai/api-fields/ for details.}
}
\value{
A data frame containing the desired data.
}
\description{
Windsor fetch Google Ads
Fetch data from google Ads via windsor.ai API
}
\examples{
\dontrun{
# api_key needs to be provided by the user.
my_googleads_data <- windsor_fetch_googleAds(api_key = "your api key",
date_preset = "last_7d",
fields = c("source", "campaign", "clicks",
           "medium", "sessions", "spend"))
}
}
